fix: Vertex interfaces moved to a separate file.
